CTS/Backend/myenv/requirements.txt
chromadb>=0.5.3
PyPDF2>=3.0.1
google-generativeai>=0.7.2
python-dotenv>=1.0.1
